Rename package from 'protocol' to 'xian_uwp' for PyPI distribution
- Renamed protocol/ folder to xian_uwp/ for better PyPI package naming
- Updated all import statements across examples and tests
- Moved test files to dedicated tests/ directory with proper __init__.py
- Updated pyproject.toml with correct package configuration
- Fixed documentation (README.md and QUICK_REFERENCE.md) to reflect new structure
- Verified all functionality still works:
  ✅ Core protocol components (models, server, client)
  ✅ CLI wallet example (create, status commands)
  ✅ Desktop and web wallet imports
  ✅ Universal DApp functionality
  ✅ All component tests pass
