+++ examples/wallets/desktop.py
@@ -3,8 +3,8 @@
 
 import flet as ft
 import threading
-from protocol.server import WalletProtocolServer
-from protocol.models import WalletType
+from xian_uwp.server import WalletProtocolServer
+from xian_uwp.models import WalletType
 
 
 class DesktopWallet:
