+++ tests/__init__.py
@@ -1,0 +1,1 @@
+# Tests for Xian Universal Wallet Protocol
