+++ tests/test_e2e.py
@@ -154,8 +154,8 @@ async def test_protocol_e2e():
 
 async def start_test_server():
     """Start a test server in the background"""
-    from protocol.server import WalletProtocolServer
-    from protocol.models import WalletType
+    from xian_uwp.server import WalletProtocolServer
+    from xian_uwp.models import WalletType
     from xian_py.wallet import Wallet
     from xian_py.xian import Xian
     import hashlib
